requirements.txt
requests
beautifulsoup4
slacker
slackclient
sqlalchemy

